+++ src/main/java/pl/plajer/villagedefense/commands/arguments/game/LeaderboardArgument.java
@@ -92,9 +92,9 @@ public class LeaderboardArgument {
       } catch (NullPointerException ex) {
         UUID current = (UUID) stats.keySet().toArray()[stats.keySet().toArray().length - 1];
         if (registry.getPlugin().getConfigPreferences().getOption(ConfigPreferences.Option.DATABASE_ENABLED)) {
-          try (Connection connection = registry.getPlugin().getMysqlDatabase().getConnection()) {
-            Statement statement = connection.createStatement();
-            ResultSet set = statement.executeQuery("SELECT name FROM playerstats WHERE UUID='" + current.toString() + "'");
+          try (Connection connection = registry.getPlugin().getMysqlDatabase().getConnection();
+               Statement statement = connection.createStatement();
+               ResultSet set = statement.executeQuery("SELECT name FROM playerstats WHERE UUID='" + current.toString() + "'")) {
             if (set.next()) {
               sender.sendMessage(formatMessage(statistic, set.getString(1), i + 1, stats.get(current)));
               continue;